--- src/types/custom.token.d.ts
@@ -1,6 +1,0 @@
-export type CustomToken = {
-  symbol: string;
-  decimals: number;
-  address: string;
-  availableBalance: number | BigNumber;
-};
