--- src/types/locked.token.distribution.d.ts
@@ -1,4 +1,0 @@
-export type LockedTokenDistribution = {
-  tx_hash: string | undefined;
-  chain_id: number;
-};
